Upvote downvote front end

+++ rails_api/app/controllers/upvotes_controller.rb
@@ -2,17 +2,21 @@ class UpvotesController < ApplicationController
   def create
     user = get_user(request.headers)
     if user
-      Upvote.find_or_create_by(user_id: user.id, link_id: params['_json'])
+      if Upvote.exists?(user_id: user.id, link_id: params['_json'])
+        render json: {message: 'You already upvoted that link'}, status: 400
+      else
+        upvote = Upvote.create(user_id: user.id, link_id: params['_json'])
+        render json: upvote.id
+      end
     else
-
+      render json: {message: 'You must be logged in to upvote'}, status: 401
     end
-    render json: {params: params, user: user}
   end
 
   def destroy
     user = get_user(request.headers)
     upvote = Upvote.find(params[:id])
-    if upvote
+    if upvote && user && upvote.user_id == user.id
       upvote.destroy
     end
   end
